Published figures checked against a measured file on every commit
Figures in the README had been maintained by hand and had drifted.
measure-counts.mjs derives them from the tree and records them in
docs/testing/measured-counts.json; --check fails the build when the two
disagree. The gate then failed on a comment, which is the wrong form of
strictness, and was corrected.

Also here: every response type completed, a counter that truncated
silently beyond two billion, two unsuccessful attempts to make CodeQL
recognise the log sanitiser — recorded as unsuccessful rather than
claimed — and the governance change that ends direct pushes to main:
CODEOWNERS, a ruleset requiring a pull request and a green verdict, and
the correction of three published claims that were not true.

+++ src/main/java/com/sm/instagram/platform/common/security/GeoLocationGdprService.java
@@ -216,7 +216,10 @@ public class GeoLocationGdprService {
         // Count travel patterns
         String pattern = TRAVEL_PREFIX + userId + ":*";
         Set<String> keys = redisTemplate.keys(pattern);
-        int travelRecords = 0;
+        // long, not int: redisTemplate.opsForList().size() returns a Long, and `int += Long` is
+        // a compound assignment with an implicit narrowing cast -- it compiles, and truncates
+        // silently past Integer.MAX_VALUE (CodeQL java/implicit-cast-in-compound-assignment).
+        long travelRecords = 0;
         
         // Defensive null check for Redis operation
         if (keys != null) {
